backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
sentence-transformers
faiss-cpu
groq
redis
python-dotenv
pydantic-settings
pymupdf
